MainProjectFolder/requirements.txt
APScheduler==3.11.2
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
defusedxml==0.7.1
diff-match-patch==20241021
Django==6.0.3
django-apscheduler==0.7.0
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-import-export==4.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.15
et_xmlfile==2.0.0
idna==3.11
inflection==0.5.1
oauthlib==3.3.1
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
psycopg2==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.2
python3-openid==3.2.0
pytz==2026.1.post1
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
tablib==3.9.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
